support modular config and systemd
Latest meta-swupdate provides a shell script wrapper to start swupdate.
This allows to put shell code snippets into the folder
/usr/lib/swupdate/conf.d to generate the parameters for swupdate at
run-time, in a very flexible way. Based on that meta-swupdate-boards
is now generic enough to work for systemd and sysv based DISTROs.

Note: For the beaglebone board swupdate 2019.11 is highly recommended.
      This version provides proper support for force_ro protection.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>