raspberry: overlayfs: Add the Install section
Add the Install section so that the overlays-etc-dirs service
can be automatically started by systemd.

Without the Install section the following warning is seen when
enabling the service manually:

systemctl enable overlays-etc-dirs

The unit files have no installation config (WantedBy=, RequiredBy=, Also=,
Alias= settings in the [Install] section, and DefaultInstance= for template
units). This means they are not meant to be enabled using systemctl.

Signed-off-by: Fabio Estevam <festevam@gmail.com>